Updated the TIH Google Analytics interface for v2 of the iOS SDK.
Added the Google Analytics v2 iOS SDK and modified the GoogleAnalytics.m file
to work correctly with the new changes. This require making GoogleAnalytics a
Singleton as it needed a iVar, which in turn means I ended up changing a lot of
files to use '[GoogleAnalytics instance]'. Also commented out the
ARGoogleAnalytics file which isn't being used, but was causing issues with no
longer having the previous Google Analytics lib.