Added all checks for access for iOS 6 EventKit operations.
User is now prompted when the app opens to allow the TIH app to access their
calendar. If they deny they are presented with a UIAlertView once, which details
how they can get that access permission backs. On iOS 5 the app should function
fine without the need for access, but this is yet to be tested as I need a real
iOS 5 device.